DELTA DEVEL15-windows-adjustable-block-size-20070810 AUTHOR jaltman@secure-endpoints.com
New registry value "BlockSize" can be used to specify an alternative
block size.  The default is 4K.  A larger blocksize will be needed if
you want to support a 6TB cache.

Also extend the service startup timeout hint to two minutes to give
the AFS client service more time to startup successfully when the
cache is really large.