DELTA DEVEL15-windows-callbacks-20070909 AUTHOR jaltman@secure-endpoints.com
when suspending only clear callbacks for servers that are in an up state.
if they were in a down state, then we couldn't send the clear callbacks
message but more importantly, we need to leave the callback info current
because the callbacks should not be cleared until the server is once again
reachable.