DELTA DEVEL15-windows-aclent-remove-extraneous-code-20071024 AUTHOR jaltman@secure-endpoints.com
Microsoft provided a dump showing cm_data.aclLRUEndp == NULL in
GetFreeACLEnt().  Couldn't find any reason why that would be true.
However, there is extraneous code that manipulates the value of
aclLRUEndp when it should be left to osi_QRemoveHT() to update
the pointer values.

Also add an additional explicit test for aclLRUEndp == NULL in
GetFreeACLEnt().