DELTA DEVEL15-windows-bplus-tree-20071103 AUTHOR jaltman@secure-endpoints.com
rename findNode to leafNode in bplus_Lookup

replace all OutputDebugString calls with osi_LogX calls

modify bestMatch to special case the return values for leaf nodes.
If an entry is above or below the values available in the leaf node
return BTLOWER or BTUPPER instead of BTERROR.

In insert and delete operations check for BTLOWER/BTUPPER and isleaf,
if true convert to either slot 0 or Max and perform the insertion.
This produces easier to read code when performing lookups.