DELTA DEVEL15-windows-dir-20071103 AUTHOR jaltman@secure-endpoints.com
Reorganize the locking for cm_BeginDirOp and cm_EndDirOp.
There are a number of locations where locks are obtained, dropped, and
reobtained.  This reorganization attempts to accomplish several things:

(1) be optimistic for the most common case so it will be fast

(2) add consistency checks after each location where locks are dropped
    and re-obtained.  If we lose a race in cm_BeginDirOp and the bplus
    tree is out of date, retry until we get to a consistent state
    that we can use.

(3) Ensure that all operations take place with the correct locks.