DELTA DEVEL15-windows-afsd-notification-20071104 AUTHOR jaltman@secure-endpoints.com
the difference between osi_assert() and osi_assertx() is whether or not
a description is specified as part of the call.  When no message is
specified afsd_notifier() is called with a NULL msgp parameter.  This
results in a null pointer dereference during the EventLog() call.

As a result, none of the information describing the assertion is written
to the log file.

This commit sets a default message for use when no other message is
provided and it replaces all calls to osi_assert with osi_assertx and
adds descriptions.