DELTA DEVEL15-windows-daemon-threads-shutdown-sync-20071223 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Windows Error Reporting received crashes caused by a failure to synchronize
the shutdown of the AFS client service with the background daemon threads.
The daemon threads (cm_Daemon, cm_BkgDaemon+, cm_IPAddrDaemon) could be
accessing data structures as they were being freed or unmapped.

Add synchronization mechanisms to signal the termination of the threads
with the shutdown of the service.