DELTA DEVEL15-rx-event-handler-20071225 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Update event_handler() for Windows.  The return code of
pthread_cond_timedwait() is not limited to 0 and -1 because 'errno'
cannot be set.  Instead of setting 'errno' the error code is returned
directly by the function as the return code.

This patch is only for debugging purposes to track the number of
success, timeout, and other errors experienced by the routine.