DELTA DEVEL15-windows-vista-network-change-20080105 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
After a network change, test to see if the desired LAN adapter to bind
the Netbios Name to has changed.  If so, stop the Netbios Network and
restart with the new LAN adapter.

This can happen on Vista after a suspend if the service wakes up before
the Loopback adapter re-initializes.