DELTA DEVEL15-cmdebug-cellservdb-20080112 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Add -cellservdb option to the cmdebug command.

This option will retrieve the volume location database information
for all cells known to the cache manager and output the data in the
format required by the CellServDB file.

This functionality is supported by the UNIX cache managers and
Windows cache managers 1.5.31 and later.