DELTA DEVEL15-windows-pioctl-path-charset-translation-20080205 AUTHOR jaltman@secure-endpoints.com LICENSE MIT FIXES 84279
Do not translate the character set of the pioctl path data.
These paths were not provided by the file system but by the end user.
Therefore, they were not converted to the OEM character set.

With this change the AFS shell context menu will be displayed for
non-ASCII file names when ANSI file names are being used as well
as the fs commands will work.