DELTA DEVEL15-windows-do-not-discard-badtickets-20080218 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
The RXKADBADTICKET error is returned when a krb5 derived token is sent
to a server that does not support them.  In a mixed cell it is possible
that some servers were not updated.  Discarding the token does not make
it possible to perform the request successfully and adversely affects
the user experience because those servers that do support the token
can no longer be accessed in an authenticated manner.  Users can't
determine what the cause is, all they see are their tokens disappearing
and there is nothing they can do about it anyway.

From now on return STATUS_NO_KERB_KEY and do not retry.