DELTA STABLE14-viced-qsort-cba-before-multibreakcallback-to-avoid-lockup-20080218 AUTHOR shadow@dementia.org LICENSE IPL10 FIXES
my code, though, this is a result of a problem reported by Chaskiel Grundman and analysis by him, Jeff Altman and myself.

I'm just checking in my implementation. In any case, the issue this addresses
is one where we can end up in makecall_waiting in rx on multiple connections
when we multibreakcallback because the lists are sorted differently and each has "pending" calls on a different connection. by sorting by index we will not
block on another caller while also holding what they're after