DELTA DEVEL15-windows-smb-writedata-crash-20080307 AUTHOR jaltman@secure-endpoints.com LICENSE MIT FIXES 88731
If an application opens a file for read only and then attempts to
write to it, we would attempt to release an scp mutex without having
obtained either the scp or the mutex within smb_WriteData

There was no check in smb_ReadData to ensure that the file was opened
for reading.  Add one.