DELTA DEVEL15-rx-mutex-interlocked-macros-20080311 AUTHOR jaltman@secure-endpoints.com LICENSE IPL10
Introduce a new set of macros that can be used to permit either mutex
based protection or Interlocked operation protection increments,
decrements, or additions.

  rx_MutexIncrement(object, mutex)
  rx_MutexAdd(object, addend, mutex)
  rx_MutexDecrement(object, mutex)
  rx_MutexAdd1Increment2(object1, addend, object2, mutex)
  rx_MutexAdd1Decrement2(object1, addend, object2, mutex)

For Windows these are implemented with the Interlocked operations
for other platforms the existing mutex is relied upon.

Only a subset of the rx_stats parameters have been transitioned at the
current time.