DELTA DEVEL15-linux-fakestat-avoid-mtpt-fillin-issue-20080415 AUTHOR rtb@pclella.cern.ch LICENSE IPL10 FIXES 93898
"It is actually the detection of a "cell" in the mount point string
which triggers the (loosely consistent) fakestat handling - it's
treated like a foreign cell and the mount point is never "completely
evaluated", i.e. logically replaced by the root directory of the
volume in question. As a result, callbacks are ignored, as they go
against the directory and not the mount point."