DELTA windows-rx-idle-timeout-20080428 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Add "IdleDeadTimeout" registry value and call rx_SetConnIdleDeadTime()
with it when creating new rx connections.   This forces the Rx stack
to return RX_CALL_TIMEOUT if the file server has returned VBUSY for
longer than the IdleDeadTimeout.

When RX_CALL_TIMEOUT is received, the cache manager will now attempt to
find another server for the object that might be less busy.