DELTA tbutc-wrong-lp-getrwindex-20080428 AUTHOR hans@MPA-Garching.MPG.DE LICENSE IPL10 FIXES 96575
The pthread butc crashes with SIGSEGV, when I restore to a volume
which already exists. This does happen with OpenAFS 1.4.7pre4 on
i386_linux26.
Looking into the code I think butc is using the wrong function
Lp_GetRwIndex (at volser/vsprocs.c:4394). It uses the one in the
library lib/libvosadmin.a, but should use the function defined
in volser/lockprocs.c.