DELTA client-idledeadtime-support-20080430 AUTHOR shadow@dementia.org LICENSE IPL10 FIXES
in the event a server is "melting down" and not responding with data to RPCs
this adds the ability to time out the calls rather than letting keepalives
effectively hang the client forever.  does not mark the server down since...
it's not down.

Side effect: adds similar tracking for authentication errors so tokens need
not be discarded if one server's clock has drifted.