DELTA rx-init-more-packets-20080529 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
It is frequently the case that rx_getAllAddr() is called before
rx_Init() or rx_InitHost().  rx_getAllAddr() obtains the list of
interfaces by using rx_GetIFInfo() which in turn computes and
allocates the number of addition rx packets.  Unfortunately,
rxi_MorePackets() relies on the existence of an initialized mutex
and the mutex is not initialized (on Windows) until the rx_InitHost()
call.  Therefore, we must delay the rxi_MorePackets() call until
after rx_InitHost() if rx_getAllAddr() is called previously.

Failure to do so results in a panic.