DELTA windows-unicode-20080626 AUTHOR asanka@secure-endpoints.com LICENSE MIT
This patch is the second stage of the Unicode conversion.  In this stage
the cache manager has been converted from using 8-bit UTF8 C strings to
16-bit UTF16 C strings in the SMB, the DNLC, the B+ tree, and other
directly related modules.  The cm_cell, cm_volume, cm_scache, cm_buf,
cm_dcache, and cm_dir modules are left 8-bit because their data is all
8-bit UTF8 since they work only on file server strings.

The SMB layer accepts 16-bit UTF16, the B+ tree and DNLC use normalized
strings as the key, and everything uses UTF8.  Efforts have been made
to minimize the number of transitions from UTF8 to UTF16 and back.

For the most part strings are tagged with clientchar_t and normchar_t
and fschar_t types in order to distinguish between the various types of
strings that are in use.

other changes include addition prototyping.