DELTA DEVEL15-windows-btree-enum-bulk-stat-20080714 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Refactor cm_TryBulkStat() to permit BulkStat RPCs to be executed
for lists of entries obtained either from the cm_dir structures
or a btree enumeration.   Add cm_BPlusDirEnumBulkStat() which
takes an enumeration object and performs a bulk status rpc on
every entry in the enumeration without advancing the enumeration
state.