DELTA DEVEL15-windows-cm-makedir-20080724 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Extend the cm_MakeDir() function signature to permit the (optional)
return of the cm_scache_t object that was created by the operation
in the same way that cm_Create() does.

Make the return of the cm_scache_t by cm_Create() optional although
it is never used that way.

In both cm_Create() and cm_MakeDir() avoid releasing the cm_scache_t
object until we are actually finished with it.