DELTA DEVEL15-windows-buf-setdirty-20080724 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
Everytime the CM_BUF_DIRTY flag is set on a cm_buf_t, the userp field
on the cm_buf_t must also be set.  Add a cm_user_t parameter to buf_SetDirty()
so that each function that calls it doesn't have to manually set the
last write user.  This improves code readability and the abstraction layering.