DELTA windows-smb-dead-vc-gc-on-head-20080801 AUTHOR jaltman@secure-endpoints.com LICENSE MIT
deltas windows-smb-dead-vc-gc-20080627 and
windows-dead-vc-cleanup-take-two-20080703 attempted to
protect against an infinite recursion when cleaning up
dead smb virtual circuits.  they failed to address the
incrementing of the vc refcount in smb_ReleaseVCInternal
when it is about to call smb_CleanupDeadVC.  If the vc
is already being cleaned, then smb_ReleaseVCInternal
should not increment the refCount and should not call
smb_CleanupDeadVC.

(This commit was committed incorrectly to the 1-5
branch first.  delta name on the head was modified
to avoid the delta name collision.)